feat: most of the required features (#1)
* feat: added staff screens

* feat: added table widget + error alert dialog

* feat: added api for get patient profile from server

* feat: adder common screen app. details & dashboard

* feat: added illness dropdown + getAppointment callback

* feat: added doctor screens

* feat: added doctor screens

* feat: added remember password

+++ lib/provider/user_account.dart
@@ -16,6 +16,8 @@ class UserAccount extends ChangeNotifier {
   static User? user;
   static String? token;
 
+  static bool savePassword = false;
+
   static _saveUser(String jwt, Map<String, dynamic> myuser) async {
     try {
       user = User.fromJson(myuser);
@@ -29,10 +31,38 @@ class UserAccount extends ChangeNotifier {
     }
   }
 
+  static _rememberUser(UserAuth user) async {
+    try {
+      await storage.write(key: 'remember', value: jsonEncode(user));
+    } on Exception catch (_) {
+      null;
+    }
+  }
+
+  static _removeRememberUser() async {
+    try {
+      await storage.delete(key: 'remember');
+    } on Exception catch (_) {
+      null;
+    }
+  }
+
   static _removeUser() async {
-    await storage.delete(key: 'jwt');
-    await storage.delete(key: 'user');
-    await storage.delete(key: 'patient');
+    try {
+      await storage.delete(key: 'jwt');
+      await storage.delete(key: 'user');
+
+      if (user!.userType == 'patient') {
+        await storage.delete(key: 'patient');
+        await storage.delete(key: 'appointment');
+      }
+
+      if (user!.userType == 'doctor') {
+        await storage.delete(key: 'doctor');
+      }
+    } on Exception catch (_) {
+      null;
+    }
 
     controller.add(false);
   }
@@ -115,6 +145,10 @@ class UserAccount extends ChangeNotifier {
         if (response.statusCode == 200) {
           _saveUser(parsed["jwt"], parsed["user"]);
 
+          final userAuth = UserAuth(email, password);
+
+          savePassword ? _rememberUser(userAuth) : _removeRememberUser();
+
           return true;
         } else {
           return parsed["error"]["message"];
@@ -130,4 +164,16 @@ class UserAccount extends ChangeNotifier {
   static Future logout() async {
     return _removeUser();
   }
+}
+
+class UserAuth {
+  final String email;
+  final String password;
+
+  UserAuth(this.email, this.password);
+
+  Map<String, dynamic> toJson() => {'email': email, 'password': password};
+
+  static UserAuth fromJson(Map<String, dynamic> json) =>
+      UserAuth(json['email'], json['password']);
 }
